eliminate frontend lint warnings
